FPS should be float

+++ motion/__main__.py
@@ -6,7 +6,7 @@ def size_tuple(size_str):
 
 if __name__ == '__main__':
 	parser = ArgumentParser(description='Process some integers.')
-	parser.add_argument('--fps', type=int, default=10)
+	parser.add_argument('--fps', type=float, default=10)
 	parser.add_argument('--size', type=size_tuple, default=(1280,720), help='The frame size in WxH')
 
 	args = parser.parse_args()
